Messages/Editor

+++ src/components/Messages.tsx
@@ -10,7 +10,6 @@ import AuthLocalStorage from "../AuthLocalStorage";
 import MessageApi from "../api/messageApi";
 import { MessageThreadModal } from './MessageThreadModal';
 import { useMessageStore } from "../stores/message.store";
-import { C } from '@fullcalendar/core/internal-common';
 
 const pageSize = 6;
 
@@ -93,21 +92,21 @@ export const Messages: React.FC = () => {
         setCurrentPage(page);
     };
 
-    const openMessageModal = (receiverName : any, recieverId?: any, ) => {
+    const openMessageModal = (receiverName : any) => {
         actions.setReceiverName(receiverName);
-        //console.log(user.UserName + " --- " + receiverName);
-        messageActions.recieveThread(user.UserName, receiverName);
+        messageActions.recieveThread(state.senderName, receiverName);
         actions.makeThreadModalVisible();
-        console.log(recieverId);
+        setContainer("Inbox");
     };
 
     const deleteMessage = async (id : any, userName : any) => {
-        await msgService.deleteMessage(id, userName);
-        await msgService.getMessages(currentPage, pageSize, container, user.NameIdentifier)
-        .then(async (response) => {
-           settotalItems(response.data.totalItems);
-           setMessages(response.data.pagedList);
-    });
+        if(window.confirm('Are you sure?')){
+            await msgService.deleteMessage(id, userName);
+            await msgService.getMessages(currentPage, pageSize, container, user.NameIdentifier)
+            .then(async (response) => {
+            settotalItems(response.data.totalItems);
+            setMessages(response.data.pagedList); });
+        }
     };
 
     const fetchData = async () => {
@@ -115,7 +114,6 @@ export const Messages: React.FC = () => {
             .then(async (response) => {
                settotalItems(response.data.totalItems);
                setMessages(response.data.pagedList);
-               console.log(messages);
         });
 
         //console.log(user);
